Adicion de interfaces para ver las actividades  On branch main  Changes to be committed: 	new file:   app/Http/Controllers/ActivitiesController.php 	new file:   app/Http/Controllers/CategoriesController.php 	modified:   app/Http/Controllers/LugaresController.php 	new file:   app/Http/Controllers/PointsController.php 	modified:   app/Models/Activities.php 	modified:   app/Models/Categories.php 	modified:   app/Models/Permisos.php 	modified:   app/Models/Points.php 	modified:   app/Models/Roles.php 	modified:   app/Models/UsersPermisos.php 	modified:   database/migrations/2021_02_14_1613316567_create_points_table.php 	modified:   database/migrations/2021_02_14_1613316664_create_activities_table.php 	modified:   database/seeders/ActivitiesSeeder.php 	modified:   database/seeders/CategoriesSeeder.php 	modified:   database/seeders/DatabaseSeeder.php 	modified:   database/seeders/PointsSeeder.php 	new file:   public/css/activities.css 	new file:   public/css/categories.css 	modified:   resources/views/lugares/index.blade.php 	new file:   resources/views/lugares/places.blade.php 	modified:   routes/web.php

+++ database/migrations/2021_02_14_1613316567_create_points_table.php
@@ -13,12 +13,12 @@ class CreatePointsTable extends Migration
             $table->bigIncrements('id');
 		    $table->text('name');
 		    $table->text('description')->nullable();
-		    $table->float('lon')->nullable();
-		    $table->float('lat')->nullable();
-		    $table->float('elevation')->nullable();
-		    $table->float('temp')->nullable();
-		    $table->float('temp_min')->nullable();
-		    $table->float('temp_max')->nullable();
+		    $table->decimal('lon', 35,30)->nullable();
+		    $table->decimal('lat', 35,30)->nullable();
+		    $table->decimal('elevation', 10,2)->nullable();
+		    $table->decimal('temp', 5,2)->nullable();
+		    $table->decimal('temp_min', 5,2)->nullable();
+		    $table->decimal('temp_max', 5,2)->nullable();
             $table->timestamps();
         });
     }
